Sheet 6 (Math - Geometry)

+++ Sheet 6 (Math - Geometry)/Q. Product .cpp
@@ -4,18 +4,18 @@ using namespace std;
 
 int main() {
 
-	long long number_1, number_2, number_3;
-	cin >> number_1 >> number_2 >> number_3;
+  long long number_1, number_2, number_3;
+  cin >> number_1 >> number_2 >> number_3;
 
-	long long max_number = max(number_1, number_2);
-	long long min_number = min(number_1, number_2);
-	long long result = 1;
+  long long max_number = max(number_1, number_2);
+  long long min_number = min(number_1, number_2);
+  long long result = 1;
 
-	for(int i = min_number; i <= max_number; i++){
-		result *= i;
-		result %= number_3;
-	}
+  for (long long i = min_number; i <= max_number; i++) {
+    result *= i;
+    result %= number_3;
+  }
 
-	cout << result << endl;
+  cout << result << endl;
 
 }
